pyproject.toml
[tool.poetry]
name = "db_compare"
version = "0.1.27"
description = "Tool to compare database schema and objects"
authors = ["sametkaradag <sametkaradag@gmail.com>"]
license = "Google"
readme = "README.md"
packages = [
    {include = "importer", from = "src/db_compare"},
    {include = "reporter", from = "src/db_compare"},
    {include = "pgcollector", from = "src/db_compare/collector"},
    {include = "oracollector", from = "src/db_compare/collector"},
    {include = "compare", from = "src/db_compare"}
]


[tool.poetry.dependencies]
python = "^3.9"
google-cloud = "0.34.0"
google-cloud-bigquery = "^3.25.0"
pyyaml = "6.0.1"
tabulate = "^0.9.0"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.32"
pandas = "^2.2.2"
oracledb = "^2.3.0"
google-cloud-secret-manager = "^2.21.1"

[tool.poetry.scripts]
pgcollector = "pgcollector.__main__:main"
oracollector = "oracollector.__main__:main"
importer = "importer.__main__:main"
reporter = "reporter.__main__:main"
compare = "compare.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
